Move .*Params classes to params package
