+++ src/main/java/api/GoblintService.java
@@ -1,6 +1,7 @@
 package api;
 
 import api.messages.*;
+import api.messages.params.*;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import org.eclipse.lsp4j.jsonrpc.services.JsonRequest;
