--- src/main/java/api/messages/ARGStateParams.java
@@ -1,15 +1,0 @@
-package api.messages;
-
-/**
- * @since 0.0.4
- */
-
-public class ARGStateParams {
-
-    private String node;
-
-    public ARGStateParams(String node) {
-        this.node = node;
-    }
-
-}
