--- src/main/java/api/messages/AnalyzeParams.java
@@ -1,11 +1,0 @@
-package api.messages;
-
-public class AnalyzeParams {
-
-    boolean reset;
-
-    public AnalyzeParams(boolean reset) {
-        this.reset = reset;
-    }
-
-}
