--- src/main/java/api/messages/NodeParams.java
@@ -1,18 +1,0 @@
-package api.messages;
-
-/**
- * @since 0.0.4
- */
-
-public class NodeParams {
-
-    private String nid;
-
-    public NodeParams() {
-    }
-
-    public NodeParams(String nid) {
-        this.nid = nid;
-    }
-
-}
