--- src/main/java/api/messages/Params.java
@@ -1,14 +1,0 @@
-package api.messages;
-
-public class Params {
-
-    private String fname;
-
-    public Params() {
-    }
-
-    public Params(String fname) {
-        this.fname = fname;
-    }
-
-}
