+++ src/main/java/api/messages/params/ARGStateParams.java
@@ -1,0 +1,15 @@
+package api.messages.params;
+
+/**
+ * @since 0.0.4
+ */
+
+public class ARGStateParams {
+
+    private String node;
+
+    public ARGStateParams(String node) {
+        this.node = node;
+    }
+
+}
