+++ src/main/java/api/messages/params/AnalyzeParams.java
@@ -1,0 +1,11 @@
+package api.messages.params;
+
+public class AnalyzeParams {
+
+    boolean reset;
+
+    public AnalyzeParams(boolean reset) {
+        this.reset = reset;
+    }
+
+}
