+++ src/main/java/api/messages/params/EvalIntQueryParams.java
@@ -1,0 +1,17 @@
+package api.messages.params;
+
+/**
+ * @since 0.0.4
+ */
+
+public class EvalIntQueryParams {
+
+    private String node;
+    private String exp;
+
+    public EvalIntQueryParams(String node, String exp) {
+        this.node = node;
+        this.exp = exp;
+    }
+
+}
