+++ src/main/java/api/messages/params/NodeParams.java
@@ -1,0 +1,18 @@
+package api.messages.params;
+
+/**
+ * @since 0.0.4
+ */
+
+public class NodeParams {
+
+    private String nid;
+
+    public NodeParams() {
+    }
+
+    public NodeParams(String nid) {
+        this.nid = nid;
+    }
+
+}
