+++ src/main/java/api/messages/params/Params.java
@@ -1,0 +1,14 @@
+package api.messages.params;
+
+public class Params {
+
+    private String fname;
+
+    public Params() {
+    }
+
+    public Params(String fname) {
+        this.fname = fname;
+    }
+
+}
